create TriggerModule
