+++ src/main/java/io/cryostat/agent/MainModule.java
@@ -37,8 +37,7 @@ import javax.net.ssl.X509TrustManager;
 import io.cryostat.agent.harvest.HarvestModule;
 import io.cryostat.agent.remote.RemoteContext;
 import io.cryostat.agent.remote.RemoteModule;
-import io.cryostat.agent.triggers.TriggerEvaluator;
-import io.cryostat.agent.triggers.TriggerParser;
+import io.cryostat.agent.triggers.TriggerModule;
 import io.cryostat.core.net.JFRConnection;
 import io.cryostat.core.net.JFRConnectionToolkit;
 import io.cryostat.core.sys.Environment;
@@ -65,6 +64,7 @@ import org.slf4j.LoggerFactory;
             ConfigModule.class,
             RemoteModule.class,
             HarvestModule.class,
+            TriggerModule.class,
         })
 public abstract class MainModule {
 
@@ -72,7 +72,6 @@ public abstract class MainModule {
     private static final int NUM_WORKER_THREADS = 3;
     private static final String JVM_ID = "JVM_ID";
     private static final String TEMPLATES_PATH = "TEMPLATES_PATH";
-    private static final String TRIGGER_SCHEDULER = "TRIGGER_SCHEDULER";
 
     @Provides
     @Singleton
@@ -239,32 +238,8 @@ public abstract class MainModule {
 
     @Provides
     @Singleton
-    @Named(TRIGGER_SCHEDULER)
-    public static ScheduledExecutorService provideTriggerScheduler() {
-        return Executors.newScheduledThreadPool(0);
-    }
-
-    @Provides
-    @Singleton
     public static FlightRecorderHelper provideFlightRecorderHelper() {
         return new FlightRecorderHelper();
-    }
-
-    @Provides
-    @Singleton
-    public static TriggerParser provideTriggerParser(FlightRecorderHelper helper) {
-        return new TriggerParser(helper);
-    }
-
-    @Provides
-    @Singleton
-    public static TriggerEvaluator provideTriggerEvaluatorFactory(
-            @Named(TRIGGER_SCHEDULER) ScheduledExecutorService scheduler,
-            TriggerParser parser,
-            FlightRecorderHelper helper,
-            @Named(ConfigModule.CRYOSTAT_AGENT_SMART_TRIGGER_EVALUATION_PERIOD_MS)
-                    long evaluationPeriodMs) {
-        return new TriggerEvaluator(scheduler, parser, helper, evaluationPeriodMs);
     }
 
     @Provides
